move data models to project data model package
