+++ backend/src/main/kotlin/com/baltic/road/path/app/model/Book.kt
@@ -1,4 +1,4 @@
-package com.baltic.road.path.app.gr.model
+package com.baltic.road.path.app.model
 
 // todo: add another required fields
 class Book {
